支持 遵从习悦 规范 和 非规范两种方式 生成代码

+++ codegen-impl/src/main/java/cn/org/rapid_framework/generator/GeneratorFacade.java
@@ -22,6 +22,9 @@ import cn.org.rapid_framework.generator.provider.page.Action;
 import cn.org.rapid_framework.generator.util.BeanHelper;
 import cn.org.rapid_framework.generator.util.GLogger;
 import cn.org.rapid_framework.generator.util.GeneratorException;
+import cn.thinkjoy.codegen.*;
+import cn.thinkjoy.codegen.GeneratorMain;
+
 /**
  * 
  * @author badqiu
@@ -174,16 +177,17 @@ public class GeneratorFacade {
 
 		public void processByAllTable(String templateRootDir,boolean isDelete) throws Exception {
 			//yq add 先拷贝managerui-biz-startup 到将来的生成目录
-			String startupDir = GeneratorProperties.getRequiredProperty("startupDir");
-			String autoGenProject = startupDir + "-autogen";
-			File autoGenDir = new File(autoGenProject);
-			if(autoGenDir.exists()){
-				//autoGenDir.delete();
-				//已存在，删除文件夹
-				Runtime.getRuntime().exec("rm -rf " + autoGenProject);
+			if(GeneratorMain.isStandard) {
+				String startupDir = GeneratorProperties.getRequiredProperty("startupDir");
+				String autoGenProject = startupDir + "-autogen";
+				File autoGenDir = new File(autoGenProject);
+				if (autoGenDir.exists()) {
+					//autoGenDir.delete();
+					//已存在，删除文件夹
+					Runtime.getRuntime().exec("rm -rf " + autoGenProject);
+				}
+				Runtime.getRuntime().exec("cp -r " + startupDir + " " + autoGenProject);
 			}
-			Runtime.getRuntime().exec("cp -r " + startupDir + " " + autoGenProject);
-
 
  			List<Table> tables = TableFactory.getInstance().getAllTables();
